separate encryption modules
