+++ src/core/writers/PDFWriter.ts
@@ -7,7 +7,8 @@ import type { PDFDict } from 'src/core/objects/PDFDict';
 import type { PDFObject } from 'src/core/objects/PDFObject';
 import type { PDFRef } from 'src/core/objects/PDFRef';
 import { PDFStream } from 'src/core/objects/PDFStream';
-import type { PDFSecurity, EncryptFn } from 'src/core/security/PDFSecurity';
+import type { PDFSecurity } from 'src/core/security/PDFSecurity';
+import type { EncryptFn } from 'src/core/security/Encryption';
 import { PDFObjectStream } from 'src/core/structures/PDFObjectStream';
 import { CharCodes } from 'src/core/syntax/CharCodes';
 import { copyStringIntoBuffer, waitForTick } from 'src/utils';
